+++ src/GuestConfiguration/GuestConfiguration/ChangeLog.md
@@ -19,6 +19,9 @@
 -->
 ## Upcoming Release
 
+## Version 0.10.8
+* Support ARC machines
+
 ## Version 0.10.7
 * Update references in .psd1 to use relative path